analytics-service/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
psycopg2-binary==2.9.9
gunicorn==20.1.0
python-dotenv==0.21.0
requests==2.28.1
boto3==1.26.50
opentelemetry-sdk==1.23.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation-flask==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-instrumentation-sqlalchemy==0.44b0
